add parameters to tuning mechanism
size: 3993
bench: 2347106

+++ src/uci.cpp
@@ -24,7 +24,7 @@ void uci() {
         if (auto value = get_if<int*>(&variant)) {
             printf("option name %s type string default %d\n", name.c_str(), **value);
         }
-        if (auto value = get_if<float*>(&variant)) {
+        if (auto value = get_if<double*>(&variant)) {
             printf("option name %s type string default %.17g\n", name.c_str(), **value);
         }
     }
@@ -64,7 +64,7 @@ void uci() {
                     if (auto value = get_if<int*>(&variant)) {
                         **value = atoi(strtok(0, " \n"));
                     }
-                    if (auto value = get_if<float*>(&variant)) {
+                    if (auto value = get_if<double*>(&variant)) {
                         **value = atof(strtok(0, " \n"));
                     }
                 }
